Change to dividing by tabstops
I don't understand the reason for the default behavior on line 92; when I edit MATLAB code I'm forced to use four spaces for indentation, and swapping in line 95 works well for me. Is the default intended for people who can use tabs?  If so, this could be improved by considering `et` vs. `noet` for users of spaces.